ai-engine/rag/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0
torch==2.2.2
transformers==4.40.1